+++ tests/test_transforms.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta, timezone
 from nose.tools import assert_equal, assert_dict_equal, assert_raises
 
 from aw_core.models import Event
-from aw_core.transform import chunk, filter_afk_events
+from aw_core.transforms import chunk, filter_afk_events
 
 
 def test_chunk():
